backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
asyncpg==0.30.0
redis==5.2.1
pydantic==2.10.4
pydantic-settings==2.7.1
PyJWT==2.10.1
